+++ docs/changes.rst
@@ -2,6 +2,11 @@
 OpenREM version history
 =======================
 
+0.9.2b1 (2019-xx-xx)
+--------------------
+* `#761`_  Imports: fixed issue in high dose alert e-mail code where week_delta may be used before assignment
+* `#746`_  Imports: enabled import of GE Elite Mini View C-arm RDSR with no template declaration
+
 0.9.1 (2019-05-16)
 ------------------
 * `#755`_  Interface: fix more static URLs to allow virtual directory web server configurations